fix tests and reorganize tests
